Rename codec to common
