+++ ethereum_history_api/oracles/src/noir/oracles/accountOracle.ts
@@ -2,7 +2,7 @@ import { type ForeignCallOutput } from '@noir-lang/noir_js';
 import { type Hex, type PublicClient } from 'viem';
 import { assert } from '../../util/assert.js';
 import { encodeAccount, encodeStateProof } from './accountOracle/encode.js';
-import { decodeAddress, decodeField } from './codec/decode.js';
+import { decodeAddress, decodeField } from './common/decode.js';
 import { NoirArguments } from './oracles.js';
 
 const GET_ACCOUNT_ARGS_COUNT = 2;
