+++ ethereum_history_api/oracles/src/noir/oracles/common/decode.ts
