+++ ethereum_history_api/oracles/src/noir/oracles/common/encode.ts
