+++ ethereum_history_api/oracles/src/noir/oracles/headerOracle/encode.ts
@@ -2,7 +2,7 @@ import { ForeignCallOutput } from '@noir-lang/noir_js';
 import { keccak256, hexToBytes } from 'viem';
 import { BlockHeader, headerToRlp } from '../../../ethereum/blockHeader.js';
 import { padArray } from '../../../util/array.js';
-import { encodeHex, encodeField } from '../codec/encode.js';
+import { encodeHex, encodeField } from '../common/encode.js';
 
 export const MAX_HEADER_RLP_SIZE = 708;
