+++ ethereum_history_api/oracles/src/noir/oracles/proofOracle.ts
@@ -2,7 +2,7 @@ import { type ForeignCallOutput } from '@noir-lang/noir_js';
 import { type Hex, type PublicClient } from 'viem';
 import { assert } from '../../util/assert.js';
 import { encodeAccount, encodeStateProof, encodeStorageProof } from './accountOracle/encode.js';
-import { decodeAddress, decodeBytes32, decodeField } from './codec/decode.js';
+import { decodeAddress, decodeBytes32, decodeField } from './common/decode.js';
 import { NoirArguments } from './oracles.js';
 
 const GET_PROOF_ARGS_COUNT = 3;
