+++ ethereum_history_api/oracles/src/script/noir_fixtures/state_proof.ts
@@ -1,6 +1,6 @@
 import { GetProofReturnType } from 'viem';
 import { encodeHexString } from '../../noir/noir_js/encode.js';
-import { encodeProof } from '../../noir/oracles/codec/encode.js';
+import { encodeProof } from '../../noir/oracles/common/encode.js';
 import { STATE_PROOF_LENGTH, encodeValue } from '../../noir/oracles/accountOracle/encode.js';
 
 export function createStateProofFixture(stateProof: GetProofReturnType): string {
