+++ ethereum_history_api/oracles/src/script/noir_fixtures/storage_proof.ts
@@ -1,6 +1,6 @@
 import { Hash } from 'viem';
 import { encodeHexString } from '../../noir/noir_js/encode.js';
-import { encodeBytes32, encodeProof } from '../../noir/oracles/codec/encode.js';
+import { encodeBytes32, encodeProof } from '../../noir/oracles/common/encode.js';
 import { STORAGE_PROOF_LENGTH } from '../../noir/oracles/accountOracle/encode.js';
 
 interface StorageProof {
